requirements.txt
https://github.com/Rapptz/discord.py/archive/rewrite.zip
antlr4-python3-runtime
